Update auth tests
